+++ planetiler-core/src/test/java/com/onthegomap/planetiler/reader/geojson/GeoJsonReaderTest.java
@@ -1,4 +1,4 @@
-package com.onthegomap.planetiler.reader;
+package com.onthegomap.planetiler.reader.geojson;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
